WIP: lots of work on time stuff! Also rustfmt-ing.
